Revert "feat: executor optimizations (#1692)"
This reverts commit 209d2e26d27a4d3236392b4bf5216e48e56b4bf3, reversing
changes made to a2dfd90abab4cf5e22acf78dd3bcbd3e77760220.

+++ crates/core/executor/src/instruction.rs
@@ -15,7 +15,7 @@ pub struct Instruction {
     /// The operation to execute.
     pub opcode: Opcode,
     /// The first operand.
-    pub op_a: u8,
+    pub op_a: u32,
     /// The second operand.
     pub op_b: u32,
     /// The third operand.
@@ -31,7 +31,7 @@ impl Instruction {
     #[must_use]
     pub const fn new(
         opcode: Opcode,
-        op_a: u8,
+        op_a: u32,
         op_b: u32,
         op_c: u32,
         imm_b: bool,
